tests/wpt/tests/css/css-masking/parsing/mask-repeat-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Masking Module Level 1: parsing mask-repeat with valid values</title>
<link rel="help" href="https://drafts.fxtf.org/css-masking-1/#the-mask-repeat">
<meta name="assert" content="mask-repeat supports the full grammar.">
<!--
<mask-repeat> = <repeat-style>#
<repeat-style> = repeat-x | repeat-y | [repeat | space | round | no-repeat]{1,2}
 -->
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("mask-repeat", "repeat-x");
test_valid_value("mask-repeat", "repeat-y");

test_valid_value("mask-repeat", "repeat");
test_valid_value("mask-repeat", "space");
test_valid_value("mask-repeat", "round");
test_valid_value("mask-repeat", "no-repeat");

test_valid_value("mask-repeat", "repeat repeat", "repeat");
test_valid_value("mask-repeat", "repeat space");
test_valid_value("mask-repeat", "repeat round");
test_valid_value("mask-repeat", "repeat no-repeat", "repeat-x");

test_valid_value("mask-repeat", "space repeat");
test_valid_value("mask-repeat", "space space", "space");
test_valid_value("mask-repeat", "space round");
test_valid_value("mask-repeat", "space no-repeat");

test_valid_value("mask-repeat", "round repeat");
test_valid_value("mask-repeat", "round space");
test_valid_value("mask-repeat", "round round", "round");
test_valid_value("mask-repeat", "round no-repeat");

test_valid_value("mask-repeat", "no-repeat repeat", "repeat-y");
test_valid_value("mask-repeat", "no-repeat space");
test_valid_value("mask-repeat", "no-repeat round");
test_valid_value("mask-repeat", "no-repeat no-repeat", "no-repeat");
</script>
</body>
</html>
